+++ src/frontend/sider/Fagsak/Behandling/sider/Behandlingsresultat/Eøs/Kompetanse/KompetanseTabellRad.tsx
@@ -11,7 +11,7 @@ import {
 import { BehandlingÅrsak, type IBehandling } from '../../../../../../../typer/behandling';
 import { KompetanseResultat, type IRestKompetanse } from '../../../../../../../typer/eøsPerioder';
 import { lagPersonLabel } from '../../../../../../../utils/formatter';
-import { StatusBarnCelleOgPeriodeCelle } from '../EøsPeriode/fellesKomponenter';
+import { StatusBarnCelleOgPeriodeCelle } from '../EøsKomponenter/EøsSkjemaKomponenter';
 
 interface IProps {
     kompetanse: IRestKompetanse;
