+++ src/frontend/sider/Fagsak/Behandling/sider/Behandlingsresultat/Eøs/UtbetaltAnnetLand/UtenlandskPeriodeBeløpTabellRadEndre.tsx
@@ -18,9 +18,9 @@ import {
     utenlandskPeriodeBeløpIntervaller,
     type IUtenlandskPeriodeBeløp,
 } from '../../../../../../../typer/eøsPerioder';
-import EøsPeriodeSkjema from '../EøsPeriode/EøsPeriodeSkjema';
-import { StyledFamilieValutavelger } from '../EøsPeriode/FamilieLandvelger';
-import { EøsPeriodeSkjemaContainer, Knapperad } from '../EøsPeriode/fellesKomponenter';
+import EøsPeriodeSkjema from '../EøsKomponenter/EøsPeriodeSkjema';
+import { EøsPeriodeSkjemaContainer, Knapperad } from '../EøsKomponenter/EøsSkjemaKomponenter';
+import { StyledFamilieValutavelger } from '../EøsKomponenter/FamilieLandvelger';
 
 const UtbetaltBeløpRad = styled.div`
     width: 32rem;
