+++ src/frontend/sider/Fagsak/Behandling/sider/Behandlingsresultat/Eøs/Valutakurs/ValutakursTabellRadEndre.tsx
@@ -15,9 +15,9 @@ import Datovelger from '../../../../../../../komponenter/Datovelger/Datovelger';
 import type { IBehandling } from '../../../../../../../typer/behandling';
 import type { OptionType } from '../../../../../../../typer/common';
 import { EøsPeriodeStatus, type IValutakurs } from '../../../../../../../typer/eøsPerioder';
-import EøsPeriodeSkjema from '../EøsPeriode/EøsPeriodeSkjema';
-import { StyledFamilieValutavelger } from '../EøsPeriode/FamilieLandvelger';
-import { EøsPeriodeSkjemaContainer, Knapperad } from '../EøsPeriode/fellesKomponenter';
+import EøsPeriodeSkjema from '../EøsKomponenter/EøsPeriodeSkjema';
+import { EøsPeriodeSkjemaContainer, Knapperad } from '../EøsKomponenter/EøsSkjemaKomponenter';
+import { StyledFamilieValutavelger } from '../EøsKomponenter/FamilieLandvelger';
 
 const ValutakursRad = styled.div`
     width: 32rem;
